Remove legacy stub usage report json
