--- core/src/test/kotlin/io/specmatic/stub/HttpStubUsageReportTest.kt
@@ -1,309 +1,0 @@
-package io.specmatic.stub
-
-import com.fasterxml.jackson.databind.ObjectMapper
-import io.specmatic.conversions.OpenApiSpecification
-import io.specmatic.core.HttpRequest
-import io.specmatic.license.core.SpecmaticProtocol
-import io.specmatic.reporter.generated.dto.stub.usage.HTTPStubUsageOperation
-import io.specmatic.reporter.generated.dto.stub.usage.SpecmaticStubUsageReport
-import io.specmatic.reporter.generated.dto.stub.usage.StubUsageEntry
-import io.specmatic.reporter.model.SpecType
-import io.specmatic.stub.report.StubEndpoint
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.AfterAll
-import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.Test
-import org.springframework.web.client.RestTemplate
-import java.io.File
-import kotlin.concurrent.thread
-
-class HttpStubUsageReportTest {
-    companion object {
-        private val stubUsageReportFile = File("./build/reports/specmatic/stub_usage_report.json")
-
-        @BeforeAll
-        @AfterAll
-        @JvmStatic
-        fun cleanUpAnyExistingReports() {
-            stubUsageReportFile.delete()
-        }
-
-        private val helloAndDataSpec = """
-openapi: 3.0.0
-info:
-  title: Sample API
-  version: 0.1.9
-paths:
-  /data:
-    get:
-      summary: hello world
-      description: test
-      responses:
-        '200':
-          description: Says hello
-          content:
-            text/plain:
-              schema:
-                type: number
-  /hello:
-    get:
-      summary: hello world
-      description: say hello
-      responses:
-        '200':
-          description: Says hello
-          content:
-            text/plain:
-              schema:
-                type: string
-  
-        """.trimIndent()
-
-        private val hello2AndData2Spec = """
-openapi: 3.0.0
-info:
-  title: Sample API
-  version: 0.1.9
-paths:
-  /data2:
-    get:
-      summary: hello world
-      description: test
-      responses:
-        '200':
-          description: Says hello
-          content:
-            text/plain:
-              schema:
-                type: number
-  /hello2:
-    get:
-      summary: hello world
-      description: say hello
-      responses:
-        '200':
-          description: Says hello
-          content:
-            text/plain:
-              schema:
-                type: string
-  
-        """.trimIndent()
-
-    }
-
-    @Test
-    fun `should determine all the api endpoints across all the specs`() {
-        val stubContract1 = OpenApiSpecification.fromYAML(helloAndDataSpec, "").toFeature()
-        val stubContract2 = OpenApiSpecification.fromYAML(hello2AndData2Spec, "").toFeature()
-
-        HttpStub(listOf(stubContract1, stubContract2)).use { stub ->
-            assertThat(stub.allEndpoints).isEqualTo(listOf(
-                StubEndpoint("/data", "GET", 200, responseContentType = "text/plain", specification = "", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-                StubEndpoint("/hello", "GET", 200, responseContentType = "text/plain", specification = "", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-                StubEndpoint("/data2", "GET", 200, responseContentType = "text/plain", specification = "", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-                StubEndpoint("/hello2", "GET", 200, responseContentType = "text/plain", specification = "", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI)
-            ))
-        }
-    }
-
-    @Test
-    fun `should generate stub usage report when stub is stopped`() {
-        val stubContract1 = OpenApiSpecification.fromYAML(
-            helloAndDataSpec,
-            "helloAndData.yaml",
-            sourceProvider = "git",
-            sourceRepository = "https://github.com/specmatic/specmatic-order-contracts.git",
-            sourceRepositoryBranch = "main",
-            specificationPath = "in/specmatic/examples/store/helloAndDataSpec.yaml"
-        ).toFeature()
-        val stubContract2 = OpenApiSpecification.fromYAML(
-            hello2AndData2Spec,
-            "hello2AndData2Spec.yaml",
-            sourceProvider = "git",
-            sourceRepository = "https://github.com/specmatic/specmatic-order-contracts.git",
-            sourceRepositoryBranch = "main",
-            specificationPath = "in/specmatic/examples/store/hello2AndData2Spec.yaml"
-        ).toFeature()
-
-        HttpStub(
-            listOf(stubContract1, stubContract2),
-            specmaticConfigSource = SpecmaticConfigSource.fromConfigFile("./specmatic.json")
-        ).use { stub ->
-            stub.client.execute(HttpRequest("GET", "/data"))
-            stub.client.execute(HttpRequest("GET", "/unknown"))
-            stub.client.execute(HttpRequest("GET", "/hello"))
-            stub.client.execute(HttpRequest("GET", "/unknown"))
-            stub.client.execute(HttpRequest("GET", "/hello2"))
-            stub.client.execute(HttpRequest("GET", "/hello2"))
-            stub.client.execute(HttpRequest("GET", "/unknown"))
-            stub.client.execute(HttpRequest("GET", "/data2"))
-            stub.client.execute(HttpRequest("GET", "/data2"))
-        }
-
-        val stubUsageReport: SpecmaticStubUsageReport = ObjectMapper().readValue(
-            stubUsageReportFile.readText(),
-            SpecmaticStubUsageReport::class.java
-        )
-
-        assertThat(stubUsageReport).isEqualTo(
-            SpecmaticStubUsageReport()
-                .withSpecmaticConfigPath(StubUsageReportTest.CONFIG_FILE_PATH)
-                .withStubUsage(
-                    listOf(
-                        StubUsageEntry()
-                            .withType("git")
-                            .withRepository("https://github.com/specmatic/specmatic-order-contracts.git")
-                            .withSpecification("in/specmatic/examples/store/helloAndDataSpec.yaml")
-                            .withBranch("main")
-                            .withServiceType("HTTP")
-                            .withSpecType("OPENAPI")
-                            .withOperations(
-                                listOf(
-                                    HTTPStubUsageOperation()
-                                        .withPath("/data")
-                                        .withMethod("GET")
-                                        .withResponseCode(200)
-                                        .withCount(1),
-                                    HTTPStubUsageOperation()
-                                        .withPath("/hello")
-                                        .withMethod("GET")
-                                        .withResponseCode(200)
-                                        .withCount(1)
-                                )
-                            ),
-                        StubUsageEntry()
-                            .withType("git")
-                            .withRepository("https://github.com/specmatic/specmatic-order-contracts.git")
-                            .withSpecification("in/specmatic/examples/store/hello2AndData2Spec.yaml")
-                            .withBranch("main")
-                            .withServiceType("HTTP")
-                            .withSpecType("OPENAPI")
-                            .withOperations(
-                                listOf(
-                                    HTTPStubUsageOperation()
-                                        .withPath("/data2")
-                                        .withMethod("GET")
-                                        .withResponseCode(200)
-                                        .withCount(2),
-                                    HTTPStubUsageOperation()
-                                        .withPath("/hello2")
-                                        .withMethod("GET")
-                                        .withResponseCode(200)
-                                        .withCount(2)
-                                )
-                            )
-                    )
-                )
-        )
-    }
-
-    @Test
-    fun `should log all successful requests when response is faked`() {
-        val contract = OpenApiSpecification.fromYAML(helloAndDataSpec, "").toFeature()
-
-        HttpStub(contract).use { stub ->
-            stub.client.execute(HttpRequest("GET", "/data"))
-            stub.client.execute(HttpRequest("GET", "/hello"))
-
-            assertThat(stub.logs).isEqualTo(listOf(
-                StubEndpoint("/data", "GET", 200, "text/plain", "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-                StubEndpoint("/hello", "GET", 200, "text/plain", "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI)
-            ))
-        }
-    }
-
-    @Test
-    fun `should log all successful requests when response is stubbed`() {
-        val contract = OpenApiSpecification.fromYAML(helloAndDataSpec, "").toFeature()
-
-        HttpStub(contract).use { stub ->
-            stub.setExpectation("""
-                {
-                    "http-request": {
-                        "method": "GET",
-                        "path": "/data"
-                    },
-                    "http-response": {
-                        "status": 200,
-                        "body": 10
-                    }
-                }
-            """.trimIndent())
-
-            val response = stub.client.execute(HttpRequest("GET", "/data"))
-            assertThat(response.body.toStringLiteral()).isEqualTo("10")
-
-            stub.client.execute(HttpRequest("GET", "/hello"))
-
-            stub.client.execute(HttpRequest("GET", "/unknown"))
-
-            assertThat(stub.logs).isEqualTo(listOf(
-                StubEndpoint("/data", "GET", 200, "text/plain", "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-                StubEndpoint("/hello", "GET", 200, "text/plain", "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI)
-            ))
-        }
-    }
-
-    @Test
-    fun `should not log unsuccessful requests`() {
-        val contract = OpenApiSpecification.fromYAML(helloAndDataSpec, "").toFeature()
-
-        HttpStub(contract).use { stub ->
-
-            stub.client.execute(HttpRequest("GET", "/data"))
-            stub.client.execute(HttpRequest("GET", "/unknown"))
-
-            assertThat(stub.logs).isEqualTo(listOf(
-                StubEndpoint("/data", "GET", 200, "text/plain", "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-            ))
-        }
-    }
-
-    @Test
-    fun `should not log unsuccessful requests in strict mode`() {
-        val contract = OpenApiSpecification.fromYAML(helloAndDataSpec, "").toFeature()
-
-        HttpStub(features = listOf(contract), strictMode = true).use { stub ->
-
-            stub.setExpectation("""
-                {
-                    "http-request": {
-                        "method": "GET",
-                        "path": "/data"
-                    },
-                    "http-response": {
-                        "status": 200,
-                        "body": 10
-                    }
-                }
-            """.trimIndent())
-
-            stub.client.execute(HttpRequest("GET", "/data"))
-            stub.client.execute(HttpRequest("GET", "/hello"))
-
-            assertThat(stub.logs).isEqualTo(listOf(
-                StubEndpoint("/data", "GET", 200, "text/plain", "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI),
-            ))
-        }
-    }
-
-    @Test
-    fun `should log all requests successfully when multiple threads make requests concurrently`() {
-        val contract = OpenApiSpecification.fromYAML(helloAndDataSpec, "").toFeature()
-
-        HttpStub(contract).use { stub ->
-
-            val threads = List(10) {
-                thread {
-                    RestTemplate().getForEntity("http://localhost:9000/hello", String::class.java)
-                }
-            }
-
-            threads.forEach { it.join() }
-
-            assertThat(stub.logs.count()).isEqualTo(10)
-            assertThat(stub.logs.all { it == StubEndpoint("/hello", "GET", 200, responseContentType = "text/plain", protocol = SpecmaticProtocol.HTTP, specType = SpecType.OPENAPI) }).isTrue
-        }
-    }
-}
